requirements.txt
aiohttp==3.9.3
python-socketio==5.11.1
redis==5.0.1
paramiko==3.4.0
requests==2.31.0
pydantic-settings==2.2.1
fastapi==0.110.0
uvicorn==0.27.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
python-dotenv==1.0.0 
pytest==7.4.3